Rename Fourier interpolation files to FourierInterpolation.h and FourierInterpolation.cpp
