+++ Code/Source/solver/FourierInterpolation.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) Stanford University, The Regents of the
 // University of California, and others. SPDX-License-Identifier: BSD-3-Clause
 
-#include "fourier_interpolation.h"
+#include "FourierInterpolation.h"
 #include "consts.h"
 
 #include <algorithm>
